+++ src/homework2/Main.java
@@ -1,4 +1,4 @@
-package java2_HW.homework2;
+package homework2;
 
 
 import java.util.Arrays;
