+++ src/homework2/MyDataArrayException.java
@@ -1,4 +1,4 @@
-package java2_HW.homework2;
+package homework2;
 
 public class MyDataArrayException extends NumberFormatException {
     public MyDataArrayException(String s) {
