--- src/java2_HW/homework1/.idea/Wall.java
@@ -1,14 +1,0 @@
-package java2_HW.homework1;
-
-public class Wall {
-    private int wallheight ;
-
-    public Wall(int wallheight) {
-        this.wallheight = wallheight;
-    }
-
-    public int getWallheight() {
-        return wallheight;
-    }
-
-}
